Now we can program the SWI!
There are a total of 18 fully programmable key slots in the MCU.
It is set up that pressing an SWI button will broadcast this intent;

tk.rabidbeaver.mcureceiver.MCU_KEY with an integer extra "KEY".

The value of KEY will be from 1 to 18 inclusive, or whatever the total
number of programmed keys is. This value will correspond to the ORDER
that the keys are programmed in.

The intent can be received using this project, which also allows actions
to be associated with the keys;

https://github.com/lbdroid/MCUKeyReceiver