Handle BT and WIFI power on suspend/resume
On suspend;
if wifi is on, flag that it needs to be resumed and turn it off.
if bt is on, flag that it needs to be resumed and turn it off.
flag that mcu is going off.

On resume;
if mcu is flagged as off, then....
if wifi is flagged to resume, turn wifi on.
if bt is flagged to resume, turn bt on.

NOTE: it will remember if the bt and wifi are on or off, and put them back in the
same state when it resumes.