Move bluetooth off to ACC off
This way the bluetooth will disconnect right when you turn off the
ignition rather than waiting potentially up to 5 minutes.

However, there should probably be some kind of override for the
ON-OFF-ON situation that happens when you start the car. For
example, if it was switched on less than, say 30 seconds ago by
ACC ON, then do a post-delayed shutoff on ACC OFF for 30 seconds,
and validate that the ACC is still off when running it.